bigsis-brain/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai>=1.0.0
typing-extensions
pydantic
asyncpg
sqlalchemy
alembic
pgvector
PyYAML
pypdf
tqdm
fpdf
psycopg2-binary
